feat: add async client
