+++ supabase/__init__.py
@@ -3,7 +3,8 @@ from postgrest import APIResponse as PostgrestAPIResponse
 from storage3.utils import StorageException
 
 from .__version__ import __version__
-from .client import Client, create_client
-from .lib.auth_client import SupabaseAuthClient
+from ._sync.auth_client import SyncSupabaseAuthClient as SupabaseAuthClient
+from ._sync.client import Client
+from ._sync.client import SyncStorageClient as SupabaseStorageClient
+from ._sync.client import create_client
 from .lib.realtime_client import SupabaseRealtimeClient
-from .lib.storage_client import SupabaseStorageClient
