+++ supabase/lib/__init__.py
@@ -1,3 +1,4 @@
-from supabase.lib import auth_client, realtime_client
+from supabase._async import auth_client
+from supabase.lib import realtime_client
 
 __all__ = ["auth_client", "realtime_client"]
